collector/requirements.txt
python-dotenv>=1.0.0
sqlalchemy>=2.0.29
speedtest-cli>=2.1.3
httpx>=0.27.0
python-dateutil>=2.9.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
psycopg2-binary>=2.9.0
